LayoutTests/imported/w3c/web-platform-tests/cors/status-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?1&origin=none&code=400
CONSOLE MESSAGE: line 55: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?1&origin=none&code=400 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?3&headers=x-custom&code=400
CONSOLE MESSAGE: line 55: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?3&headers=x-custom&code=400 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?2&origin=none&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?4&headers=x-custom&code=400&preflight=200 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?5&headers=x-custom&preflight=400 due to access control checks.
The returned status code in different scenarios

   allowed  preflight  response  | status |
   -------  ---------  --------  | ------ |
 1      no          x       400  |      0 |
 2      no        200         x  |      0 |
 3     yes          x       400  |    400 |
 4     yes        200       400  |    400 |
 5     yes        400         x  |      0 |

PASS 1. CORS disallowed, response status 400. 
PASS 2. CORS disallowed, preflight status 200. 
FAIL 3. CORS allowed, response status 400. assert_unreached: error event Reached unreachable code
FAIL 4. CORS allowed, preflight status 200, response status 400. assert_unreached: error event Reached unreachable code
PASS 5. CORS allowed, preflight status 400. 
